# Adding most of scipy.linalg.special_matrices.