… docs/source/reference/linalg.rst