This adds 17 out of 20 of the functions in `scipy.linalg.special_matrices`. The only missing ones are currently implemented in scipy with `scipy.special.comb` which is currently not implemented in cupy (and would also run into integer overflow issues on the GPU quickly as well). Additionally, the `tri` function actually just calls the already efficiently implemented `cupy.tri` function.

Will be adding tests shortly, however, what is the earliest still-supported scipy version? These functions were added later:
* 0.8.0 - `circulant`, `hadamard`, `leslie`, `companion`
* 0.10.0 - `hilbert`
* 0.14.0 - `dft`
* 1.3.0 - `fiedler`, `fiedler_companion`
* 1.5.0 - `convolution_matrix`