@coderforlife 2020-07-21
Looking through the tests, the requirement "scipy>=" is used at earliest with 0.19.0 so I added testing rules for only the final 3 special matrix functions.