@grlee77 2020-07-21
I don't see a specific SciPy version listed as supported in the docs, but I think scipy 0.18 is the oldest version tested on CI:
https://github.com/chainer/chainer-test/blob/master/run_multi_test.py#L22