@kmaehashi 2020-07-22
Could you split the PR for each function?

Regarding SciPy version support, we discussed and planning to only support 1.3+. (#3631)