@coderforlife 2020-07-22
@kmaehashi I can split it up a bit, but some of them are related to each other so I would probably still keep them together (like `tri`/`tril`/triu` and `toeplitz`/`circulant`/`hankel`). However, currently the tests are currently a bit intertwined.

How should I proceed? Do I make 15 different branches and submit 15 PRs right now for all of them, or do I do it one PR at a time and can reuse the same branch? Will the merging handle it just fine if they are done as 15 separate branches? It seems like it will think they should all be the "first" function in the file.