@asi1024 2020-08-09
Could you update `docs/source/reference/linalg.rst`?