@coderforlife 2020-08-10
Using a custom an `ElementwiseKernel` for `hadamard` has great results (old is the version with a loop and new is the `ElementwiseKernel` version):

`n` | scipy (µs) | old (µs) | new (µs)
---|---|---|---
256 | 151 | 306 | 12.9
512 | 517 | 339 | 15.6
1024 | 3600 | 391 | 25.5
2048 | 24600 | 521 | 65.4
4096 | 120 ms | 994 | 224
8192 | 513 ms | 2880 | 861
16384 | 2.08 s | 10.4 ms | 3.4 ms

The cupy function is now faster than scipy for all values of `n` except 1 (but previously it was only faster `n>=512`).