@coderforlife 2020-08-19
Fixed documentation issues.